небольшой рефакторинг модели, реализовано сохранение файлов по актуальному пути

+++ CertificateManagementSystem.Services/FileService.cs
@@ -1,4 +1,5 @@
 ﻿using CertificateManagementSystem.Data;
+using CertificateManagementSystem.Data.Models;
 using Microsoft.Extensions.Configuration;
 using System.IO;
 
@@ -14,19 +15,44 @@ namespace CertificateManagementSystem.Services
         }
 
         // Создать файл документа
-        public void CreateFile(string sourceFilePath, ref string destinationFilePath)
+        public void CreateFile(string sourceFilePath, string destinationFilePath)
         {
             var documentsFolderPath = _configuration.GetSection("Paths").GetSection("DocumentsFolder").Value;
             destinationFilePath = Path.Combine(documentsFolderPath, destinationFilePath);
 
-            var folder = Path.GetDirectoryName(destinationFilePath);
+            if (File.Exists(destinationFilePath))
+                return;
 
-            // Директория не существует?
-            if (!Directory.Exists(folder))
-                Directory.CreateDirectory(folder);
-            
             File.Copy(sourceFilePath, destinationFilePath);
         }
 
+        public string GetRealFilePath(FileModel file)
+        {
+            var documentsFolderPath = _configuration.GetSection("Paths").GetSection("DocumentsFolder").Value;
+            var filePath = Path.Combine(documentsFolderPath, file.Path);
+
+            var fileFolder = Path.GetDirectoryName(filePath);
+            var fileName = Path.GetFileNameWithoutExtension(filePath);
+            var extension = Path.GetExtension(filePath);
+            var i = 0;
+
+            // Проверить существет ли файл с таким именем
+            while (File.Exists(filePath))
+            {
+                var fileInfo = new FileInfo(filePath);
+
+                if (file.Size == fileInfo.Length)
+                    return filePath.Replace(documentsFolderPath, "");
+
+                var newFileName = fileName + $"_({i++})";
+                filePath = Path.Combine(fileFolder, newFileName + extension);
+            }
+
+            // Директория не существует?
+            if (!Directory.Exists(fileFolder))
+                Directory.CreateDirectory(fileFolder);
+
+            return filePath.Replace(documentsFolderPath, "");
+        }
     }
 }
